Added new class for excel handling
