+++ main/main.csproj
@@ -7,6 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="ExcelDataReader" Version="3.6.0" />
+    <PackageReference Include="ExcelNumberFormat" Version="1.1.0" />
   </ItemGroup>
 
 </Project>
